updated Spaces roles and members
added members and described the roles more verbose.
removed the info that "Personal" is actutally a space as well. this is very technical information but leads to confusion, because there is no explicit manager within personal. lets just call it the place where private/personal files should be. 
we once decided to user Space as a proper name (own name) to make clear, that this is dedicated concept. therefor, we decided to wirte Space with a capital S. updated that as well here.